+++ library/src/actions/guard/guard.ts
@@ -5,49 +5,60 @@ import type {
 } from '../../types/index.ts';
 import { _addIssue } from '../../utils/index.ts';
 
-type BaseGuard<TInput> = (
+/**
+ * Guard function type.
+ *
+ * @beta
+ */
+export type GuardFunction<TInput> = (
   input: TInput
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
 ) => input is any;
 
+/**
+ * Infer guard output type.
+ *
+ * @beta
+ */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-type InferGuarded<TGuard extends BaseGuard<any>> = TGuard extends (
+export type InferGuardOutput<TGuard extends GuardFunction<any>> =
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  input: any
-) => input is infer TOutput
-  ? TOutput
-  : unknown;
+  TGuard extends (input: any) => input is infer TOutput ? TOutput : unknown;
 
 /**
  * Guard issue interface.
+ *
+ * @beta
  */
-export interface GuardIssue<TInput, TGuard extends BaseGuard<TInput>>
+export interface GuardIssue<TInput, TGuard extends GuardFunction<TInput>>
   extends BaseIssue<TInput> {
   /**
    * The issue kind.
    */
   readonly kind: 'transformation';
   /**
-   * The validation type.
+   * The issue type.
    */
   readonly type: 'guard';
   /**
-   * The validation requirement.
+   * The guard function.
    */
   readonly requirement: TGuard;
 }
 
 /**
  * Guard action interface.
+ *
+ * @beta
  */
 export interface GuardAction<
   TInput,
-  TGuard extends BaseGuard<TInput>,
+  TGuard extends GuardFunction<TInput>,
   TMessage extends ErrorMessage<GuardIssue<TInput, TGuard>> | undefined,
 > extends BaseTransformation<
     TInput,
     // intersect in case guard is actually wider
-    TInput & InferGuarded<TGuard>,
+    TInput & InferGuardOutput<TGuard>,
     GuardIssue<TInput, TGuard>
   > {
   /**
@@ -67,43 +78,50 @@ export interface GuardAction<
    */
   readonly message: TMessage;
 }
+
 /**
- * Creates a guard validation action.
+ * Creates a guard transformation action.
  *
  * @param requirement The guard function.
  *
  * @returns A guard action.
+ *
+ * @beta
  */
 // known input from pipe
-export function guard<TInput, const TGuard extends BaseGuard<TInput>>(
+export function guard<TInput, const TGuard extends GuardFunction<TInput>>(
   requirement: TGuard
 ): GuardAction<TInput, TGuard, undefined>;
 
 /**
- * Creates a guard validation action.
+ * Creates a guard transformation action.
  *
  * @param requirement The guard function.
  *
  * @returns A guard action.
+ *
+ * @beta
  */
 // unknown input, e.g. standalone
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function guard<const TGuard extends BaseGuard<any>>(
+export function guard<const TGuard extends GuardFunction<any>>(
   requirement: TGuard
 ): GuardAction<Parameters<TGuard>[0], TGuard, undefined>;
 
 /**
- * Creates a guard validation action.
+ * Creates a guard transformation action.
  *
  * @param requirement The guard function.
  * @param message The error message.
  *
  * @returns A guard action.
+ *
+ * @beta
  */
 // known input from pipe
 export function guard<
   TInput,
-  const TGuard extends BaseGuard<TInput>,
+  const TGuard extends GuardFunction<TInput>,
   const TMessage extends ErrorMessage<GuardIssue<TInput, TGuard>> | undefined,
 >(
   requirement: TGuard,
@@ -111,17 +129,19 @@ export function guard<
 ): GuardAction<TInput, TGuard, TMessage>;
 
 /**
- * Creates a guard validation action.
+ * Creates a guard transformation action.
  *
  * @param requirement The guard function.
  * @param message The error message.
  *
  * @returns A guard action.
+ *
+ * @beta
  */
 // unknown input, e.g. standalone
 export function guard<
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const TGuard extends BaseGuard<any>,
+  const TGuard extends GuardFunction<any>,
   const TMessage extends
     | ErrorMessage<GuardIssue<Parameters<TGuard>[0], TGuard>>
     | undefined,
@@ -132,12 +152,12 @@ export function guard<
 
 // @__NO_SIDE_EFFECTS__
 export function guard(
-  requirement: BaseGuard<unknown>,
-  message?: ErrorMessage<GuardIssue<unknown, BaseGuard<unknown>>>
+  requirement: GuardFunction<unknown>,
+  message?: ErrorMessage<GuardIssue<unknown, GuardFunction<unknown>>>
 ): GuardAction<
   unknown,
-  BaseGuard<unknown>,
-  ErrorMessage<GuardIssue<unknown, BaseGuard<unknown>>> | undefined
+  GuardFunction<unknown>,
+  ErrorMessage<GuardIssue<unknown, GuardFunction<unknown>>> | undefined
 > {
   return {
     kind: 'transformation',